Fix gl not working and prepare some good method

+++ src/models/vector_2d.cpp
@@ -2,7 +2,7 @@
 
 Vector2D::Vector2D() {}
 
-Vector2D::Vector2D(int x, int y)
+Vector2D::Vector2D(float x, float y)
 {
     this->x = x;
     this->y = y;
