+++ src/models/vector_2d.h
@@ -1,9 +1,9 @@
 class Vector2D
 {
 public:
-    int x;
-    int y;
+    float x;
+    float y;
 
     Vector2D();
-    Vector2D(int x, int y);
+    Vector2D(float x, float y);
 };
